Fixed most compiler warnings -Wall -Wextra
repro:
  CFLAGS="-Wall -Wextra -Wunused-parameter -Wc++-compat" ./configure && make

which we use for perl, and libyaml is now included in cperl.
Tested with gcc-5 and clang-3.7

There are still a tons of format warnings (%d on 64bit) in
example-deconstructor.c which I skipped.

+++ tests/example-reformatter.c
@@ -120,42 +120,42 @@ parser_error:
 
         case YAML_READER_ERROR:
             if (parser.problem_value != -1) {
-                fprintf(stderr, "Reader error: %s: #%X at %zd\n", parser.problem,
-                        parser.problem_value, parser.problem_offset);
+                fprintf(stderr, "Reader error: %s: #%X at %ld\n", parser.problem,
+                        parser.problem_value, (long)parser.problem_offset);
             }
             else {
-                fprintf(stderr, "Reader error: %s at %zd\n", parser.problem,
-                        parser.problem_offset);
+                fprintf(stderr, "Reader error: %s at %ld\n", parser.problem,
+                        (long)parser.problem_offset);
             }
             break;
 
         case YAML_SCANNER_ERROR:
             if (parser.context) {
-                fprintf(stderr, "Scanner error: %s at line %lu, column %lu\n"
-                        "%s at line %lu, column %lu\n", parser.context,
-                        parser.context_mark.line+1, parser.context_mark.column+1,
-                        parser.problem, parser.problem_mark.line+1,
-                        parser.problem_mark.column+1);
+                fprintf(stderr, "Scanner error: %s at line %d, column %d\n"
+                        "%s at line %d, column %d\n", parser.context,
+                        (int)parser.context_mark.line+1, (int)parser.context_mark.column+1,
+                        parser.problem, (int)parser.problem_mark.line+1,
+                        (int)parser.problem_mark.column+1);
             }
             else {
-                fprintf(stderr, "Scanner error: %s at line %lu, column %lu\n",
-                        parser.problem, parser.problem_mark.line+1,
-                        parser.problem_mark.column+1);
+                fprintf(stderr, "Scanner error: %s at line %d, column %d\n",
+                        parser.problem, (int)parser.problem_mark.line+1,
+                        (int)parser.problem_mark.column+1);
             }
             break;
 
         case YAML_PARSER_ERROR:
             if (parser.context) {
-                fprintf(stderr, "Parser error: %s at line %lu, column %lu\n"
-                        "%s at line %lu, column %lu\n", parser.context,
-                        parser.context_mark.line+1, parser.context_mark.column+1,
-                        parser.problem, parser.problem_mark.line+1,
-                        parser.problem_mark.column+1);
+                fprintf(stderr, "Parser error: %s at line %d, column %d\n"
+                        "%s at line %d, column %d\n", parser.context,
+                        (int)parser.context_mark.line+1, (int)parser.context_mark.column+1,
+                        parser.problem, (int)parser.problem_mark.line+1,
+                        (int)parser.problem_mark.column+1);
             }
             else {
-                fprintf(stderr, "Parser error: %s at line %lu, column %lu\n",
-                        parser.problem, parser.problem_mark.line+1,
-                        parser.problem_mark.column+1);
+                fprintf(stderr, "Parser error: %s at line %d, column %d\n",
+                        parser.problem, (int)parser.problem_mark.line+1,
+                        (int)parser.problem_mark.column+1);
             }
             break;
